Move get_allocated_memory_units to mod
